Exception added

+++ db/db_user.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm.session import Session
 from schema import user_base
 from db.models import db_user
 from db.hash import hash_user
+from fastapi import HTTPException, status
 
 
 def create_user(db: Session, request: user_base):
@@ -21,7 +22,11 @@ def get_all(db: Session):
 
 
 def get_user(db: Session, id_: int):
-    return db.query(db_user).filter(db_user.id == id_).first()
+    user = db.query(db_user).filter(db_user.id == id_)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id_} not found')
+    return user
 
 
 def update_user(db: Session, id_: int, request: user_base):
@@ -37,6 +42,9 @@ def update_user(db: Session, id_: int, request: user_base):
 
 def delete_user(db: Session, id_: int):
     user = db.query(db_user).filter(db_user.id == id_).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f'User with id {id_} not found')
     db.delete(user)
     db.commit()
     return "ok"
